+++ lib/src/state_manager.dart
@@ -12,12 +12,12 @@ import '../simple_managers.dart';
 ///
 class StateManager extends CloneableFile {
   static const className = 'StateManager';
-  static const version = '4';
+  static const version = '5';
   static const String _saveKey = "map";
 
   FlagManager fm = FlagManager();
   IndexManager im = IndexManager();
-  SelectionManager sem = SelectionManager();
+  TagSelectionManager tsm = TagSelectionManager();
   MultiFlagManager mfm = MultiFlagManager();
   MultiIndexManager mim = MultiIndexManager();
   MultiSelectionManager msem = MultiSelectionManager();
@@ -38,8 +38,9 @@ class StateManager extends CloneableFile {
         fm = FlagManager.fromDict(src[_saveKey]["flag_manager"]);
       } else if (i == "index_manager") {
         im = IndexManager.fromDict(src[_saveKey]["index_manager"]);
-      } else if (i == "selection_manager") {
-        sem = SelectionManager.fromDict(src[_saveKey]["selection_manager"]);
+      } else if (i == "tag_selection_manager") {
+        tsm = TagSelectionManager.fromDict(
+            src[_saveKey]["tag_selection_manager"]);
       } else if (i == "multi_flag_manager") {
         mfm = MultiFlagManager.fromDict(src[_saveKey]["multi_flag_manager"]);
       } else if (i == "multi_index_manager") {
@@ -51,6 +52,10 @@ class StateManager extends CloneableFile {
         tfm = TextFieldManager.fromDict(src[_saveKey]["textfield_manager"]);
       } else if (i == "value_manager") {
         vm = ValueManager.fromDict(src[_saveKey]["value_manager"]);
+      }
+      // old version (under v6.0.1)
+      else if (i == "selection_manager") {
+        tsm = TagSelectionManager.fromDict(src[_saveKey]["selection_manager"]);
       }
     }
   }
@@ -79,7 +84,7 @@ class StateManager extends CloneableFile {
     Map<String, Map<String, dynamic>> mMap = {
       "flag_manager": fm.toDict(nonSaveKeys: nonSaveKeys),
       "index_manager": im.toDict(nonSaveKeys: nonSaveKeys),
-      "selection_manager": sem.toDict(nonSaveKeys: nonSaveKeys),
+      "tag_selection_manager": tsm.toDict(nonSaveKeys: nonSaveKeys),
       "multi_flag_manager": mfm.toDict(nonSaveKeys: nonSaveKeys),
       "multi_index_manager": mim.toDict(nonSaveKeys: nonSaveKeys),
       "multi_selection_manager": msem.toDict(nonSaveKeys: nonSaveKeys),
@@ -106,7 +111,7 @@ class StateManager extends CloneableFile {
     if (other is StateManager) {
       if (fm != other.fm ||
           im != other.im ||
-          sem != other.sem ||
+          tsm != other.tsm ||
           mfm != other.mfm ||
           mim != other.mim ||
           msem != other.msem ||
@@ -122,7 +127,7 @@ class StateManager extends CloneableFile {
 
   @override
   int get hashCode {
-    List<Object> objects = [fm, im, sem, mfm, mim, msem, tfm, vm];
+    List<Object> objects = [fm, im, tsm, mfm, mim, msem, tfm, vm];
     return Object.hashAll(objects);
   }
 }